# specs/001-job-search-mvp/contracts/openapi.yml
openapi: 3.1.0
info:
  title: Job Enhancer API
  description: Job Search Assistant MVP — FastAPI backend
  version: 1.0.0
paths:
  /v1/jobs/:
    get:
      tags:
      - Jobs
      summary: Search Jobs
      description: Search jobs across Adzuna and JSearch, deduplicated and paginated.
      operationId: search_jobs_v1_jobs__get
      security:
      - HTTPBearer: []
      parameters:
      - name: q
        in: query
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 255
          description: Search query
          title: Q
        description: Search query
      - name: location
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            maxLength: 255
          - type: 'null'
          title: Location
      - name: remote_only
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Remote Only
      - name: salary_min
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
            minimum: 0
          - type: 'null'
          title: Salary Min
      - name: salary_max
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
            minimum: 0
          - type: 'null'
          title: Salary Max
      - name: experience
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            pattern: ^(entry|mid|senior)$
          - type: 'null'
          title: Experience
      - name: job_type
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Job Type
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 50
          minimum: 1
          default: 20
          title: Page Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobSearchResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/jobs/{job_id}:
    get:
      tags:
      - Jobs
      summary: Get Job
      description: Fetch a single job listing by ID.
      operationId: get_job_v1_jobs__job_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Job Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobListingSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/saved-jobs/:
    get:
      tags:
      - SavedJobs
      summary: List Saved Jobs
      operationId: list_saved_jobs_v1_saved_jobs__get
      security:
      - HTTPBearer: []
      parameters:
      - name: collection_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Collection Id
      - name: pipeline_stage_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Pipeline Stage Id
      - name: is_archived
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Is Archived
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SavedJobSchema'
                title: Response List Saved Jobs V1 Saved Jobs  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - SavedJobs
      summary: Save Job
      operationId: save_job_v1_saved_jobs__post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SavedJobCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavedJobSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/saved-jobs/manual:
    post:
      tags:
      - SavedJobs
      summary: Save Manual Job
      description: Add a job found on an external site by URL + details (FR-004a).
      operationId: save_manual_job_v1_saved_jobs_manual_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManualJobCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavedJobSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /v1/saved-jobs/{saved_job_id}:
    patch:
      tags:
      - SavedJobs
      summary: Update Saved Job
      operationId: update_saved_job_v1_saved_jobs__saved_job_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: saved_job_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Saved Job Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SavedJobUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavedJobSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - SavedJobs
      summary: Delete Saved Job
      operationId: delete_saved_job_v1_saved_jobs__saved_job_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: saved_job_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Saved Job Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v1/collections/:
    get:
      tags:
      - Collections
      summary: List Collections
      operationId: list_collections_v1_collections__get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/CollectionSchema'
                type: array
                title: Response List Collections V1 Collections  Get
      security:
      - HTTPBearer: []
    post:
      tags:
      - Collections
      summary: Create Collection
      operationId: create_collection_v1_collections__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /v1/collections/{collection_id}:
    get:
      tags:
      - Collections
      summary: Get Collection
      operationId: get_collection_v1_collections__collection_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: collection_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Collection Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - Collections
      summary: Update Collection
      operationId: update_collection_v1_collections__collection_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: collection_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Collection Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Collections
      summary: Delete Collection
      operationId: delete_collection_v1_collections__collection_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: collection_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Collection Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v1/admin/stats:
    get:
      tags:
      - Admin
      summary: Get Stats
      operationId: get_stats_v1_admin_stats_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlatformStats'
      security:
      - HTTPBearer: []
...
  /v1/admin/users:
    get:
      tags:
      - Admin
      summary: List Users
      operationId: list_users_v1_admin_users_get
      security:
      - HTTPBearer: []
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 25
          title: Page Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserProfile'
                title: Response List Users V1 Admin Users Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    CollectionCreate:
      properties:
        name:
          type: string
          maxLength: 100
          minLength: 1
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        color:
          anyOf:
          - type: string
            pattern: ^#[0-9A-Fa-f]{6}$
          - type: 'null'
          title: Color
      type: object
      required:
      - name
      title: CollectionCreate
    CollectionSchema:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        user_id:
          type: string
          format: uuid
          title: User Id
        name:
          type: string
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        color:
          anyOf:
          - type: string
          - type: 'null'
          title: Color
        sort_order:
          type: integer
          title: Sort Order
        is_default:
          type: boolean
          title: Is Default
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - user_id
      - name
      - description
      - color
      - sort_order
      - is_default
      - created_at
      - updated_at
      title: CollectionSchema
    CollectionUpdate:
      properties:
        name:
          anyOf:
          - type: string
            maxLength: 100
            minLength: 1
          - type: 'null'
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        color:
          anyOf:
          - type: string
            pattern: ^#[0-9A-Fa-f]{6}$
          - type: 'null'
          title: Color
      type: object
      title: CollectionUpdate
    DailySignup:
      properties:
        date:
          type: string
          title: Date
        count:
          type: integer
          title: Count
      type: object
      required:
      - date
      - count
      title: DailySignup
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    JobListingSchema:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        external_id:
          type: string
          title: External Id
        source:
          type: string
          title: Source
        title:
          type: string
          title: Title
        company:
          type: string
          title: Company
        location:
          type: string
          title: Location
        is_remote:
          type: boolean
          title: Is Remote
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        salary_min:
          anyOf:
          - type: integer
          - type: 'null'
          title: Salary Min
        salary_max:
          anyOf:
          - type: integer
          - type: 'null'
          title: Salary Max
        currency:
          anyOf:
          - type: string
          - type: 'null'
          title: Currency
        job_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Job Type
        apply_url:
          type: string
          title: Apply Url
        posted_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Posted At
        is_expired:
          type: boolean
          title: Is Expired
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - external_id
      - source
      - title
      - company
      - location
      - is_remote
      - description
      - salary_min
      - salary_max
      - currency
      - job_type
      - apply_url
      - posted_at
      - is_expired
      - created_at
      title: JobListingSchema
    JobSearchResponse:
      properties:
        results:
          items:
            $ref: '#/components/schemas/JobListingSchema'
          type: array
          title: Results
        meta:
          $ref: '#/components/schemas/PaginatedMeta'
      type: object
      required:
      - results
      - meta
      title: JobSearchResponse
    ManualJobCreate:
      properties:
        url:
          type: string
          maxLength: 2000
          minLength: 8
          title: Url
        title:
          type: string
          maxLength: 500
          minLength: 1
          title: Title
        company:
          type: string
          maxLength: 255
          minLength: 1
          title: Company
        location:
          type: string
          maxLength: 255
          title: Location
          default: Not specified
        is_remote:
          type: boolean
          title: Is Remote
          default: false
        collection_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Collection Id
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
      type: object
      required:
      - url
      - title
      - company
      title: ManualJobCreate
      description: A job the user found on an external site (FR-004a).
...
    PaginatedMeta:
      properties:
        page:
          type: integer
          title: Page
        page_size:
          type: integer
          title: Page Size
        total:
          type: integer
          title: Total
        total_pages:
          type: integer
          title: Total Pages
      type: object
      required:
      - page
      - page_size
      - total
      - total_pages
      title: PaginatedMeta
...
    PlatformStats:
      properties:
        total_users:
          type: integer
          title: Total Users
        active_7d:
          type: integer
          title: Active 7D
        active_30d:
          type: integer
          title: Active 30D
        new_7d:
          type: integer
          title: New 7D
        signups_by_day:
          items:
            $ref: '#/components/schemas/DailySignup'
          type: array
          title: Signups By Day
      type: object
      required:
      - total_users
      - active_7d
      - active_30d
      - new_7d
      - signups_by_day
      title: PlatformStats
...
    SavedJobCreate:
      properties:
        job_listing_id:
          type: string
          format: uuid
          title: Job Listing Id
        collection_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Collection Id
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
      type: object
      required:
      - job_listing_id
      title: SavedJobCreate
    SavedJobSchema:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        user_id:
          type: string
          format: uuid
          title: User Id
        job_listing_id:
          type: string
          format: uuid
          title: Job Listing Id
        collection_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Collection Id
        pipeline_stage_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Pipeline Stage Id
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
        applied_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Applied At
        last_stage_change:
          type: string
          format: date-time
          title: Last Stage Change
        is_archived:
          type: boolean
          title: Is Archived
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        job_listing:
          $ref: '#/components/schemas/JobListingSchema'
      type: object
      required:
      - id
      - user_id
      - job_listing_id
      - collection_id
      - pipeline_stage_id
      - notes
      - applied_at
      - last_stage_change
      - is_archived
      - created_at
      - updated_at
      - job_listing
      title: SavedJobSchema
    SavedJobUpdate:
      properties:
        collection_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Collection Id
        pipeline_stage_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Pipeline Stage Id
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
        applied_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Applied At
        is_archived:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Archived
      type: object
      title: SavedJobUpdate
...
    UserProfile:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        email:
          type: string
          format: email
          title: Email
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        image:
          anyOf:
          - type: string
          - type: 'null'
          title: Image
        role:
          type: string
          title: Role
        follow_up_days:
          type: integer
          title: Follow Up Days
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - email
      - role
      - follow_up_days
      - created_at
      title: UserProfile
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
servers:
- url: http://localhost:8000
  description: Local development
- url: https://job-enhancer-api.onrender.com
  description: Production (Render)